requirements.txt
asgiref==3.9.0
Django==4.2.23
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.14.1
